Update drain template to properly manage spaces
The current template will sometimes result in a a drain getting a message with an extra space as the first character.  This fix ensure an extra space doesn't get propagated to the drain.  It is the way the rsyslog drain template recommends doing it: http://www.rsyslog.com/doc/v8-stable/configuration/templates.html#standard-template-for-forwarding-to-a-remote-host-rfc3164-mode