validate all rsyslog config are valid
When adding a custom rule, we validate that that entire rsyslog config
is valid. If it is not, we remove the custom rule and continue the
deployment. In compromise of discovering that the custom rule fails the
operator can look for the message 'Custom rule configuration invalid'
within the `/var/vcap/sys/logs/` for a failure message.

[#153385067]

Signed-off-by: Jesse Alford <jalford@pivotal.io>