Ability to use Mutual TLS with syslog endpoints
In Mutual TLS, both the client and server provide certificates for the
other party to validate. This is useful for restricting which servers
can input to a syslog endpoint.

This commit adds new properties to the `syslog_forwarded` job. These
expose rsyslog's out-of-the-box support for Mutual TLS.

This commit also adds acceptance tests, for which it was necessary to
also add Mutual TLS support to the `syslog_storer` job.