Increase simultaneous tests to 3
- Also moves the upload-stemcell action from the pipeline task to the
  script run by the task. (This commit should be paired with the
  corresponding pipeline change.)

Experimentation has determined that the Directors provisioned by the
Toolsmiths CF Deployment-creating software set things up so that the
Directors run up to three tasks simultaneously. So, we're increasing the
number of Ginkgo nodes from 2 to 3 to take advantage of that, as well as
making it configurable, so we can change it in the future.

Signed-off-by: Ben Fuller <benjaminf@vmware.com>